add svg_concat script. reprovision. remove unneeded functions.

+++ setup.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
-"""Setup module for provisioner pkg."""
+"""Setup module for text_img_creator pkg."""
 from setuptools import setup, find_packages
 import os
 import sys
@@ -7,13 +7,13 @@ import shutil
 import stat
 from setuptools.command.develop import develop
 from setuptools.command.install import install
+from abstract_requires import requires
 
 pkg_name = "text_img_creator"
 parent_dir = os.path.dirname(os.path.realpath(__file__))
 data_src_dir = pkg_name + "_data"
 config_src_dir = pkg_name + "_config"
 defaults_location = os.path.join(pkg_name, "defaults")
-requires = list(())
 
 # Create scripts list
 script_dir = os.path.join(parent_dir, "bin")
@@ -25,11 +25,6 @@ try:
 except FileNotFoundError:
     pass
 
-pkg_has_config = not True
-if pkg_has_config:
-    requires.append("provision_py_proj")
-
-
 def copy_pkg_files():
     """Copy config and data files."""
     from appdirs import user_data_dir, user_config_dir
@@ -40,26 +35,27 @@ def copy_pkg_files():
     ]
     for d, t in pkg_dirs_to_copy:
         t = os.path.join(t, d)
-        d = os.path.join(pkg_name, d)
+        d = os.path.join(os.path.join(parent_dir, pkg_name), d)
 
-        shutil.rmtree(t, ignore_errors=True)
-        shutil.copytree(d, t)
+        if os.path.exists(d):
+            shutil.rmtree(t, ignore_errors=True)
+            shutil.copytree(d, t)
 
-        user = os.environ["SUDO_USER"]
+            user = os.environ["SUDO_USER"]
 
-        for root, dirs, files in os.walk(t):
-            shutil.chown(root, user=user, group=user)
-            os.chmod(root, stat.S_IRWXU)
+            for root, dirs, files in os.walk(t):
+                shutil.chown(root, user=user, group=user)
+                os.chmod(root, stat.S_IRWXU)
 
-            for d in dirs:
-                d_path = os.path.join(root, d)
-                shutil.chown(d_path, user=user, group=user)
-                os.chmod(d_path, stat.S_IRWXU)
+                for d in dirs:
+                    d_path = os.path.join(root, d)
+                    shutil.chown(d_path, user=user, group=user)
+                    os.chmod(d_path, stat.S_IRWXU)
 
-            for f in files:
-                f_path = os.path.join(root, f)
-                shutil.chown(f_path, user=user, group=user)
-                os.chmod(f_path, stat.S_IRUSR | stat.S_IWUSR)
+                for f in files:
+                    f_path = os.path.join(root, f)
+                    shutil.chown(f_path, user=user, group=user)
+                    os.chmod(f_path, stat.S_IRUSR | stat.S_IWUSR)
 
 
 def reset():
@@ -74,10 +70,10 @@ def setuptools_setup():
     setup(
         name="text_img_creator",
         version="0.1",
-        description="",
+        description="default description",
         url="default url",
-        author="Kevin Wolf",
-        author_email="kevinuwolf@gmail.com",
+        author="default author",
+        author_email="default author",
         license="gplv3.txt",
         packages=find_packages(),
         scripts=scripts,
@@ -94,8 +90,7 @@ def main():
         reset()
     else:
         setuptools_setup()
-        if pkg_has_config:
-            copy_pkg_files()
+        copy_pkg_files()
 
 if __name__ == "__main__":
     main()
